+++ examples/native-eth-gas-sponsorship/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @renegade-fi/gas-sponsorship-example
 
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @renegade-fi/core@0.8.0
+  - @renegade-fi/node@0.6.0
+
 ## 0.0.23
 
 ### Patch Changes